# apps/intg-api/openapi-definition.yml
openapi: 3.0.3
info:
  title: OGCIO Integrator API
  description: API for OGCIO Integrator Service
  version: 0.1.0
...
paths:
...
  /api/v1/journeys/:
    get:
      tags: &a1
        - Journeys
      parameters:
        - schema:
            default: 0
            minimum: 0
            type: number
          in: query
          name: offset
          required: false
        - schema:
            default: 10
            minimum: 5
            maximum: 50
            multipleOf: 5
            type: number
          in: query
          name: limit
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        title:
                          type: string
                        userId:
                          type: string
                        organizationId:
                          type: string
                        status:
                          anyOf:
                            - type: string
                              enum:
                                - active
                            - type: string
                              enum:
                                - draft
                        initialStepId:
                          type: string
                        createdAt:
                          type: string
                        updatedAt:
                          type: string
                        userName:
                          type: string
                      required:
                        - id
                        - title
                        - userId
                        - organizationId
                        - status
                        - initialStepId
                        - createdAt
                        - updatedAt
                        - userName
                  metadata:
                    type: object
                    properties:
                      links:
                        type: object
                        properties:
                          self:
                            type: object
                            properties:
                              href:
                                type: string
                          next:
                            type: object
                            properties:
                              href:
                                type: string
                          prev:
                            type: object
                            properties:
                              href:
                                type: string
                          first:
                            type: object
                            properties:
                              href:
                                type: string
                          last:
                            type: object
                            properties:
                              href:
                                type: string
                          pages:
                            type: object
                            additionalProperties:
                              type: object
                              properties:
                                href:
                                  type: string
                        required:
                          - self
                          - first
                          - last
                          - pages
                      totalCount:
                        type: number
                required:
                  - data
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  detail:
                    type: string
                  requestId:
                    type: string
                  name:
                    type: string
                  validation: {}
                  validationContext:
                    type: string
                required:
                  - code
                  - detail
                  - requestId
                  - name
    post:
      tags: *a1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                organizationId:
                  type: string
                userId:
                  type: string
              required:
                - title
                - organizationId
                - userId
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                    required:
                      - id
                  metadata:
                    type: object
                    properties:
                      links:
                        type: object
                        properties:
                          self:
                            type: object
                            properties:
                              href:
                                type: string
                          next:
                            type: object
                            properties:
                              href:
                                type: string
                          prev:
                            type: object
                            properties:
                              href:
                                type: string
                          first:
                            type: object
                            properties:
                              href:
                                type: string
                          last:
                            type: object
                            properties:
                              href:
                                type: string
                          pages:
                            type: object
                            additionalProperties:
                              type: object
                              properties:
                                href:
                                  type: string
                        required:
                          - self
                          - first
                          - last
                          - pages
                      totalCount:
                        type: number
                required:
                  - data
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  detail:
                    type: string
                  requestId:
                    type: string
                  name:
                    type: string
                  validation: {}
                  validationContext:
                    type: string
                required:
                  - code
                  - detail
                  - requestId
                  - name
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  detail:
                    type: string
                  requestId:
                    type: string
                  name:
                    type: string
                  validation: {}
                  validationContext:
                    type: string
                required:
                  - code
                  - detail
                  - requestId
                  - name
...
  /api/v1/journeys/{journeyId}/public-info:
    get:
      tags: *a1
      parameters:
        - schema:
            type: string
          in: path
          name: journeyId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      title:
                        type: string
                      userId:
                        type: string
                      organizationId:
                        type: string
                      status:
                        anyOf:
                          - type: string
                            enum:
                              - active
                          - type: string
                            enum:
                              - draft
                      initialStepId:
                        type: string
                      createdAt:
                        type: string
                      updatedAt:
                        type: string
                    required:
                      - id
                      - title
                      - userId
                      - organizationId
                      - status
                      - initialStepId
                      - createdAt
                      - updatedAt
                  metadata:
                    type: object
                    properties:
                      links:
                        type: object
                        properties:
                          self:
                            type: object
                            properties:
                              href:
                                type: string
                          next:
                            type: object
                            properties:
                              href:
                                type: string
                          prev:
                            type: object
                            properties:
                              href:
                                type: string
                          first:
                            type: object
                            properties:
                              href:
                                type: string
                          last:
                            type: object
                            properties:
                              href:
                                type: string
                          pages:
                            type: object
                            additionalProperties:
                              type: object
                              properties:
                                href:
                                  type: string
                        required:
                          - self
                          - first
                          - last
                          - pages
                      totalCount:
                        type: number
                required:
                  - data
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  detail:
                    type: string
                  requestId:
                    type: string
                  name:
                    type: string
                  validation: {}
                  validationContext:
                    type: string
                required:
                  - code
                  - detail
                  - requestId
                  - name
  /api/v1/journeys/{journeyId}/schema:
    get:
      tags: *a1
      parameters:
        - schema:
            type: string
          in: path
          name: journeyId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      journeyId:
                        type: string
                      jounrneyTitle:
                        type: string
                      steps:
                        type: array
                        items:
                          type: object
                          properties:
                            stepId:
                              type: string
                            type:
                              anyOf:
                                - type: string
                                  enum:
                                    - form
                                - type: string
                                  enum:
                                    - payment
                                - type: string
                                  enum:
                                    - messaging
                                - type: string
                                  enum:
                                    - complete
                            resourceId:
                              type: string
                            stepSchema: {}
                          required:
                            - stepId
                            - type
                            - resourceId
                      stepConnections:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            sourceStepId:
                              type: string
                            destinationStepId:
                              type: string
                          required:
                            - id
                            - sourceStepId
                            - destinationStepId
                    required:
                      - journeyId
                      - jounrneyTitle
                      - steps
                      - stepConnections
                  metadata:
                    type: object
                    properties:
                      links:
                        type: object
                        properties:
                          self:
                            type: object
                            properties:
                              href:
                                type: string
                          next:
                            type: object
                            properties:
                              href:
                                type: string
                          prev:
                            type: object
                            properties:
                              href:
                                type: string
                          first:
                            type: object
                            properties:
                              href:
                                type: string
                          last:
                            type: object
                            properties:
                              href:
                                type: string
                          pages:
                            type: object
                            additionalProperties:
                              type: object
                              properties:
                                href:
                                  type: string
                        required:
                          - self
                          - first
                          - last
                          - pages
                      totalCount:
                        type: number
                required:
                  - data
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  detail:
                    type: string
                  requestId:
                    type: string
                  name:
                    type: string
                  validation: {}
                  validationContext:
                    type: string
                required:
                  - code
                  - detail
                  - requestId
                  - name
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  detail:
                    type: string
                  requestId:
                    type: string
                  name:
                    type: string
                  validation: {}
                  validationContext:
                    type: string
                required:
                  - code
                  - detail
                  - requestId
                  - name
...
  /api/v1/executor/runs/journeys/{journeyId}:
    get:
      tags: *a4
      parameters:
        - schema:
            type: string
          in: path
          name: journeyId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        userId:
                          type: string
                        journeyId:
                          type: string
                        status:
                          anyOf:
                            - type: string
                              enum:
                                - pending
                            - type: string
                              enum:
                                - failed
                            - type: string
                              enum:
                                - completed
                        createdAt:
                          type: string
                        updatedAt:
                          type: string
                      required:
                        - id
                        - userId
                        - journeyId
                        - status
                        - createdAt
                        - updatedAt
                  metadata:
                    type: object
                    properties:
                      links:
                        type: object
                        properties:
                          self:
                            type: object
                            properties:
                              href:
                                type: string
                          next:
                            type: object
                            properties:
                              href:
                                type: string
                          prev:
                            type: object
                            properties:
                              href:
                                type: string
                          first:
                            type: object
                            properties:
                              href:
                                type: string
                          last:
                            type: object
                            properties:
                              href:
                                type: string
                          pages:
                            type: object
                            additionalProperties:
                              type: object
                              properties:
                                href:
                                  type: string
                        required:
                          - self
                          - first
                          - last
                          - pages
                      totalCount:
                        type: number
                required:
                  - data
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  detail:
                    type: string
                  requestId:
                    type: string
                  name:
                    type: string
                  validation: {}
                  validationContext:
                    type: string
                required:
                  - code
                  - detail
                  - requestId
                  - name
...